+++ src/parser/attr.ts
@@ -1,6 +1,6 @@
-import type {NSP} from "../index.js";
+import type {NSP} from "../../index.js";
 
-import {parseText} from "./parse-text.js";
+import {parseText} from "./text.js";
 
 const LF = (indent: number) => (+indent ? "\n" + " ".repeat(indent) : "\n");
